requirements.txt
scrapy==2.0
marshmallow==3.5.2
asyncio-nats-streaming==0.4.0
python-dotenv== 0.14.0
